refactor: 유저 정보 목 데이터 위치 수정
